+++ config/dev.exs
@@ -15,4 +15,5 @@ config :pigeon, PigeonTest.APNS,
 config :pigeon, PigeonTest.FCM,
   adapter: Pigeon.FCM,
   auth: PigeonTest.Goth,
-  project_id: System.get_env("FCM_PROJECT")
+  project_id: System.get_env("FCM_PROJECT"),
+  ping_period: 60_000